requirements.txt
aiohttp==3.7.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
boto3==1.17.18
botocore==1.20.18
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
django-phonenumber-field==5.0.0
gunicorn==20.0.4
idna==2.10
ipinfo==3.0.0
ipinfo-django==1.1.0
jmespath==0.10.0
multidict==5.1.0
phonenumbers==8.12.18
Pillow==8.1.1
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
twilio==6.53.0
typing-extensions==3.7.4.3
urllib3==1.26.3
whitenoise==5.2.0
xlwt==1.3.0
yarl==1.6.3
aiohttp==3.7.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
boto3==1.17.18
botocore==1.20.18
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
django-phonenumber-field==5.0.0
gunicorn==20.0.4
idna==2.10
ipinfo==3.0.0
ipinfo-django==1.1.0
jmespath==0.10.0
multidict==5.1.0
mysqlclient==2.0.3
phonenumbers==8.12.18
Pillow==8.1.1
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
twilio==6.53.0
typing-extensions==3.7.4.3
urllib3==1.26.3
whitenoise==5.2.0
xlwt==1.3.0
yarl==1.6.3
aiohttp==3.7.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
boto3==1.17.18
botocore==1.20.18
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
django-phonenumber-field==5.0.0
gunicorn==20.0.4
idna==2.10
ipinfo==3.0.0
ipinfo-django==1.1.0
jmespath==0.10.0
multidict==5.1.0
mysqlclient==2.0.3
phonenumbers==8.12.18
Pillow==8.1.1
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
twilio==6.53.0
typing-extensions==3.7.4.3
urllib3==1.26.3
whitenoise==5.2.0
xlwt==1.3.0
yarl==1.6.3
